Langraph_Exploration/paper_recommendation_agent/requirements.txt
transformers>=4.35.0
torch>=2.0.0
datasets>=2.14.0
peft>=0.7.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
ollama>=0.1.0
streamlit>=1.28.0
requests>=2.31.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
huggingface-hub>=0.19.0